Refactored folder structure for ristretto255 (#364)
