+++ frost-ristretto255/src/keys/dkg.rs
@@ -85,3 +85,16 @@ pub fn part3(
 ) -> Result<(KeyPackage, PublicKeyPackage), Error> {
     frost::keys::dkg::part3(round2_secret_package, round1_packages, round2_packages)
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::Ristretto255Sha512;
+    use rand::thread_rng;
+
+    #[test]
+    fn check_sign_with_dkg() {
+        let rng = thread_rng();
+
+        frost_core::tests::check_sign_with_dkg::<Ristretto255Sha512, _>(rng);
+    }
+}
